chore: move gradle dependency export in a separate package
