+++ rockcraft-gradle/src/main/java/com/canonical/rockcraft/gradle/dependencies/PomDependencyReader.java
@@ -11,7 +11,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.canonical.rockcraft.gradle;
+package com.canonical.rockcraft.gradle.dependencies;
 
 import org.apache.commons.text.StringSubstitutor;
 import org.apache.maven.model.Model;
@@ -117,10 +117,12 @@ public class PomDependencyReader {
 
     private static  StringSubstitutor createPropertyReplacer(Model mavenModel) {
         HashMap<String, String> replacements = new HashMap<>();
-
         for(String propertyName : mavenModel.getProperties().stringPropertyNames()) {
             replacements.put(propertyName, mavenModel.getProperties().getProperty(propertyName));
         }
+        replacements.put("project.version", mavenModel.getVersion());
+        replacements.put("project.groupId", mavenModel.getGroupId());
+        replacements.put("project.artifactId", mavenModel.getArtifactId());
 
         return new StringSubstitutor(replacements);
     }
